refactor: update type imports to use new schema types and remove deprecated files
