+++ src/lib/schemas/create_join_code.schema.ts
