+++ src/routes/(app)/onboarding/+page.svelte
@@ -3,7 +3,7 @@
 	import { Section } from 'flowbite-svelte-blocks';
 	import { Field, Control, FieldErrors } from 'formsnap';
 	import { zodClient } from 'sveltekit-superforms/adapters';
-	import { schema } from './schema.js';
+	import { schema } from '../../../lib/schemas/onboarding.schema.js';
 	import { DateInput } from 'date-picker-svelte';
 	import { FloatingLabelInput, Input, Label, Button, Modal } from 'flowbite-svelte';
 	let { data } = $props();
