+++ src/routes/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { Snippet } from 'svelte';
 
-	import '../app.css';
+	import '$css';
 	let { children } = $props<{ children: Snippet }>();
 </script>
