+++ src/routes/home/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { Snippet } from 'svelte';
 	import Navbar from '$lib/ui/components/Navbar.svelte';
-	import type { User } from '$lib/server/types';
+	import type { User } from '$lib/schema-types';
 	import Footer from '$lib/ui/components/Footer.svelte';
 	import '../../app.css';
 	let { children, data } = $props<{ data: { user: User | null }; children: Snippet }>();
