+++ svelte.config.js
@@ -3,7 +3,15 @@ import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 const config = {
 	preprocess: vitePreprocess(),
 	kit: {
-		adapter: adapter()
+		adapter: adapter(),
+		alias: {
+			$route: 'src/route',
+			$css: 'src/app.css',
+			$schemas: 'src/lib/schemas',
+			$server: 'src/lib/server',
+			$ui: 'src/lib/ui',
+			$schema_types: 'src/lib/schema_types.ts'
+		}
 	}
 	/*
 	compilerOptions: {
